добавляет абстрактный класс и контроллер
