+++ src/components/cards-container.js
@@ -1,0 +1,15 @@
+import AbstractComponent from './abstract-component';
+
+const createCardsFilmTemplate = () => {
+
+  return (
+    `<div class="films-list__container">
+    </div>
+  `);
+};
+
+export default class CardsContainer extends AbstractComponent {
+  getTemplate() {
+    return createCardsFilmTemplate();
+  }
+}
