--- src/components/comments.js
@@ -1,65 +1,0 @@
-import {formatDateTime} from '../utils';
-
-const createCommentTemplate = (comment) => {
-  const {textComment, emoji, author, data} = comment;
-  return (
-    `<li class="film-details__comment">
-      <span class="film-details__comment-emoji">
-        <img src="${emoji}" alt="emoji-smile" width="55" height="55">
-      </span>
-      <div>
-        <p class="film-details__comment-text">${textComment}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${author}</span>
-          <span class="film-details__comment-day">${formatDateTime(data, `number`, `divider`, `number`)}</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-    </li>`);
-};
-
-const createCommentsTemplate = (commentsArray) => {
-  const commentMarkup = commentsArray.map((it) => createCommentTemplate(it)).join(`\n`);
-
-  return (
-    `<section class="film-details__comments-wrap">
-      <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsArray.length}</span></h3>
-
-      <ul class="film-details__comments-list">
-        ${commentMarkup}
-      </ul>
-
-      <div class="film-details__new-comment">
-        <div for="add-emoji" class="film-details__add-emoji-label"></div>
-
-        <label class="film-details__comment-label">
-          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-        </label>
-
-        <div class="film-details__emoji-list">
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-          <label class="film-details__emoji-label" for="emoji-smile">
-            <img src="./images/emoji/smile.png" alt="emoji" width="30" height="30">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-          <label class="film-details__emoji-label" for="emoji-sleeping">
-            <img src="./images/emoji/sleeping.png" alt="emoji" width="30" height="30">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-          <label class="film-details__emoji-label" for="emoji-puke">
-            <img src="./images/emoji/puke.png" alt="emoji" width="30" height="30">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-          <label class="film-details__emoji-label" for="emoji-angry">
-            <img src="./images/emoji/angry.png" alt="emoji" width="30" height="30">
-          </label>
-        </div>
-      </div>
-    </section>`);
-};
-
-
-export {createCommentsTemplate};
