+++ src/components/detailted-film.js
@@ -1,5 +1,66 @@
-import {createCommentsTemplate} from './comments';
-import {formatTime, formatDate, createElement} from '../utils';
+import {formatTime, formatDate, formatDateTime} from '../utils/common';
+import AbstractComponent from './abstract-component';
+
+const createCommentTemplate = (comment) => {
+  const {textComment, emoji, author, data} = comment;
+  return (
+    `<li class="film-details__comment">
+      <span class="film-details__comment-emoji">
+        <img src="${emoji}" alt="emoji-smile" width="55" height="55">
+      </span>
+      <div>
+        <p class="film-details__comment-text">${textComment}</p>
+        <p class="film-details__comment-info">
+          <span class="film-details__comment-author">${author}</span>
+          <span class="film-details__comment-day">${formatDateTime(data, `number`, `divider`, `number`)}</span>
+          <button class="film-details__comment-delete">Delete</button>
+        </p>
+      </div>
+    </li>`);
+};
+
+const createCommentsTemplate = (commentsArray) => {
+  const commentMarkup = commentsArray.map((it) => createCommentTemplate(it)).join(`\n`);
+
+  return (
+    `<section class="film-details__comments-wrap">
+      <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsArray.length}</span></h3>
+
+      <ul class="film-details__comments-list">
+        ${commentMarkup}
+      </ul>
+
+      <div class="film-details__new-comment">
+        <div for="add-emoji" class="film-details__add-emoji-label"></div>
+
+        <label class="film-details__comment-label">
+          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+        </label>
+
+        <div class="film-details__emoji-list">
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+          <label class="film-details__emoji-label" for="emoji-smile">
+            <img src="./images/emoji/smile.png" alt="emoji" width="30" height="30">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+          <label class="film-details__emoji-label" for="emoji-sleeping">
+            <img src="./images/emoji/sleeping.png" alt="emoji" width="30" height="30">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+          <label class="film-details__emoji-label" for="emoji-puke">
+            <img src="./images/emoji/puke.png" alt="emoji" width="30" height="30">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+          <label class="film-details__emoji-label" for="emoji-angry">
+            <img src="./images/emoji/angry.png" alt="emoji" width="30" height="30">
+          </label>
+        </div>
+      </div>
+    </section>`);
+};
 
 const createDetailedFilmTemplate = (film) => {
   const {age, title, originalTitle, rating, director, writers, actors, year, duration, country, genres, poster, description, commentsArray, isWatchlist, isWatched, isFavorite} = film;
@@ -71,7 +132,6 @@ const createDetailedFilmTemplate = (film) => {
             </div>
           </div>
 
-
           <section class="film-details__controls">
             <input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist" ${isWatchlist ? `checked` : ``}>
             <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>
@@ -91,25 +151,17 @@ const createDetailedFilmTemplate = (film) => {
     </section>`);
 };
 
-export default class DetaltedCard {
+export default class DetaltedCard extends AbstractComponent {
   constructor(card) {
+    super();
     this._card = card;
-    this._elements = null;
   }
 
   getTemplate() {
     return createDetailedFilmTemplate(this._card);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setClickCloseButtonHandler(handler) {
+    this.getElement().querySelector(`.film-details__close-btn`).addEventListener(`click`, handler);
   }
 }
