+++ src/components/show-button.js
@@ -1,27 +1,15 @@
-import {createElement} from '../utils';
+import AbstractComponent from './abstract-component';
 
 const createShowButtonTemplate = () => {
   return (`<button class="films-list__show-more">Show more</button>`);
 };
 
-export default class ShowButton {
-  constructor() {
-    this._element = null;
-  }
-
+export default class ShowButton extends AbstractComponent {
   getTemplate() {
     return createShowButtonTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setShowMoreButtonClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
   }
 }
