+++ src/components/sorts.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component";
 
 const createSortTemplate = (name, isActive) => {
   return (`<li><a href="#" class="sort__button ${isActive ? `sort__button--active` : ``}">Sort by ${name}</a></li>`);
@@ -12,25 +12,13 @@ const createSortsTemplate = (sorts) => {
     </ul>`);
 };
 
-export default class Sorts {
+export default class Sorts extends AbstractComponent {
   constructor(sorts) {
+    super();
     this._sorts = sorts;
-    this._element = null;
   }
 
   getTemplate() {
     return createSortsTemplate(this._sorts);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
